PresoakAndHeavySoil (ReSharper Code Cleanup following SDE merge.)

+++ src/EntityFramework/Core/EntityModel/SchemaObjectModel/ByteFacetDescriptionElement.cs
@@ -1,16 +1,12 @@
-using System;
-using System.Collections.Generic;
-using System.Text;
-using System.Data.Entity.Core.Metadata.Edm;
-using System.Xml;
-using System.Diagnostics;
-
 namespace System.Data.Entity.Core.EntityModel.SchemaObjectModel
 {
+    using System.Data.Entity.Core.Metadata.Edm;
+    using System.Xml;
+
     internal sealed class ByteFacetDescriptionElement : FacetDescriptionElement
     {
         public ByteFacetDescriptionElement(TypeElement type, string name)
-        :base(type, name)
+            : base(type, name)
         {
         }
 
@@ -31,7 +27,7 @@ namespace System.Data.Entity.Core.EntityModel.SchemaObjectModel
             byte value = 0;
             if (HandleByteAttribute(reader, ref value))
             {
-                DefaultValue = (Byte)value;
+                DefaultValue = value;
             }
         }
     }
